+++ eslint-plugin-sermonforge/lib/rules/no-direct-database.js
@@ -21,7 +21,7 @@ const SPINE_ONLY_NAMES = new Set([
   'createSeries', 'updateSeries', 'deleteSeries',
   'getSermonsBySeries',
   'getSectionsBySeries', 'createSection', 'updateSection', 'deleteSection',
-  'loadTourSermon', 'removeTourSermon',
+  'loadSampleSermon',
   'transitionState', 'applyMutation', 'persistMutation',
 ]);